Fix known bugs: bank recon Int64, keybox Last3M sort, unit toggle InjectFrom
- RptBankRecon: fix Int64 casting error (TxId int -> Int64)
- ItemSearch: add Last3M (NotMapped) and PrimaryImageUrl (NotMapped) to prevent FromSql errors
- ItemRepository: populate Last3M from Items table for keybox cache
- Item_ListActiveForKeybox SP: sort by Last3M DESC
- TempSalesService: fix InjectFrom skipping private-setter fields (Unit, OrdQty, etc.)
  in both Update() and UpdateUnit() methods
- Sales_Inject SP: add ISNULL for CartLineType and IsSystemManaged

+++ KLS.Models/Reports/Sales/RptBankRecon.cs
@@ -7,7 +7,7 @@ namespace KLS.Models.Reports
         [Key]
         public int RowId { get; set; }
 
-        public int? TxId { get; set; }
+        public Int64? TxId { get; set; }
 
         public DateOnly? TxDate { get; set; }
